requirements.txt
flask
PyGithub
google-generativeai
python-dotenv